+++ src/main/java/in/koreatech/koin/domain/user/controller/UserApi.java
@@ -12,26 +12,26 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import in.koreatech.koin._common.auth.Auth;
 import in.koreatech.koin.domain.user.dto.AuthResponse;
-import in.koreatech.koin.domain.user.dto.CheckEmailDuplicationRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckEmailDuplicationRequest;
 import in.koreatech.koin.domain.user.dto.GeneralUserRegisterRequest;
-import in.koreatech.koin.domain.user.dto.CheckNicknameDuplicationRequest;
-import in.koreatech.koin.domain.user.dto.CheckPhoneDuplicationRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckNicknameDuplicationRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckPhoneDuplicationRequest;
 import in.koreatech.koin.domain.user.dto.UserAccessTokenRequest;
 import in.koreatech.koin.domain.user.dto.UserLoginRequest;
 import in.koreatech.koin.domain.user.dto.UserLoginResponse;
-import in.koreatech.koin.domain.user.dto.CheckUserPasswordRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckUserPasswordRequest;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshRequest;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshResponse;
-import in.koreatech.koin.domain.user.dto.verification.ExistsByPhoneRequest;
-import in.koreatech.koin.domain.user.dto.verification.ExistsByUserIdRequest;
-import in.koreatech.koin.domain.user.dto.verification.ExistsByEmailRequest;
-import in.koreatech.koin.domain.user.dto.verification.MatchUserIdWithEmailRequest;
-import in.koreatech.koin.domain.user.dto.verification.MatchUserIdWithPhoneNumberRequest;
-import in.koreatech.koin.domain.user.dto.verification.FindIdByEmailRequest;
-import in.koreatech.koin.domain.user.dto.verification.ResetPasswordByEmailRequest;
-import in.koreatech.koin.domain.user.dto.verification.FindIdResponse;
-import in.koreatech.koin.domain.user.dto.verification.FindIdBySmsRequest;
-import in.koreatech.koin.domain.user.dto.verification.ResetPasswordBySmsRequest;
+import in.koreatech.koin.domain.user.dto.validation.ExistsByPhoneRequest;
+import in.koreatech.koin.domain.user.dto.validation.ExistsByUserIdRequest;
+import in.koreatech.koin.domain.user.dto.validation.ExistsByEmailRequest;
+import in.koreatech.koin.domain.user.dto.validation.MatchUserIdWithEmailRequest;
+import in.koreatech.koin.domain.user.dto.validation.MatchUserIdWithPhoneNumberRequest;
+import in.koreatech.koin.domain.user.dto.FindIdByEmailRequest;
+import in.koreatech.koin.domain.user.dto.ResetPasswordByEmailRequest;
+import in.koreatech.koin.domain.user.dto.FindIdResponse;
+import in.koreatech.koin.domain.user.dto.FindIdBySmsRequest;
+import in.koreatech.koin.domain.user.dto.ResetPasswordBySmsRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
