+++ src/main/java/in/koreatech/koin/domain/user/dto/validation/CheckUserPasswordRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.user.dto;
+package in.koreatech.koin.domain.user.dto.validation;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
