+++ src/main/java/in/koreatech/koin/domain/user/dto/validation/ExistsByPhoneRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.user.dto.verification;
+package in.koreatech.koin.domain.user.dto.validation;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
@@ -10,7 +10,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record FindIdBySmsRequest(
+public record ExistsByPhoneRequest(
     @Schema(description = "전화번호", example = "01012345678", requiredMode = REQUIRED)
     @NotBlank(message = "전화번호는 필수입니다.")
     @Pattern(regexp = "^010[0-9]{8}$", message = "올바른 전화번호 형식이 아닙니다.")
