--- src/main/java/in/koreatech/koin/domain/user/dto/verification/FindIdByEmailRequest.java
@@ -1,20 +1,0 @@
-package in.koreatech.koin.domain.user.dto.verification;
-
-import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-@JsonNaming(value = SnakeCaseStrategy.class)
-public record FindIdByEmailRequest(
-    @Schema(description = "이메일", example = "user@example.com", requiredMode = REQUIRED)
-    @NotBlank(message = "이메일은 필수입니다.")
-    @Email(message = "올바른 이메일 형식이 아닙니다.")
-    String email
-) {
-
-}
