+++ src/main/java/in/koreatech/koin/domain/user/service/UserValidationService.java
@@ -11,10 +11,10 @@ import in.koreatech.koin._common.auth.exception.AuthorizationException;
 import in.koreatech.koin._common.exception.custom.KoinIllegalArgumentException;
 import in.koreatech.koin.domain.student.model.redis.UnAuthenticatedStudentInfo;
 import in.koreatech.koin.domain.student.repository.StudentRedisRepository;
-import in.koreatech.koin.domain.user.dto.CheckEmailDuplicationRequest;
-import in.koreatech.koin.domain.user.dto.CheckNicknameDuplicationRequest;
-import in.koreatech.koin.domain.user.dto.CheckPhoneDuplicationRequest;
-import in.koreatech.koin.domain.user.dto.CheckUserPasswordRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckEmailDuplicationRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckNicknameDuplicationRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckPhoneDuplicationRequest;
+import in.koreatech.koin.domain.user.dto.validation.CheckUserPasswordRequest;
 import in.koreatech.koin.domain.user.exception.DuplicationNicknameException;
 import in.koreatech.koin.domain.user.exception.DuplicationPhoneNumberException;
 import in.koreatech.koin.domain.user.exception.UserNotFoundException;
